checkout flow updates
