+++ packages/checkout/src/views/PaymentSelection/PayWithCrypto/index.tsx
@@ -1,68 +1,37 @@
-import { Box, Button, Text, Scroll, Spinner } from '@0xsequence/design-system'
-import {
-  useBalances,
-  useContractInfo,
-  useSwapPrices,
-  useSwapQuote,
-  compareAddress,
-  TRANSACTION_CONFIRMATIONS_DEFAULT,
-  sendTransactions,
-  SwapPricesWithCurrencyInfo,
-  NATIVE_TOKEN_ADDRESS_0X
-} from '@0xsequence/kit'
+import { Box, Text, Scroll, Spinner } from '@0xsequence/design-system'
+import { useBalances, useContractInfo, useSwapPrices, compareAddress, NATIVE_TOKEN_ADDRESS_0X } from '@0xsequence/kit'
 import { findSupportedNetwork } from '@0xsequence/network'
-import { useState, useEffect, Fragment } from 'react'
-import { encodeFunctionData, formatUnits, Hex, zeroAddress } from 'viem'
-import { usePublicClient, useWalletClient, useReadContract, useAccount } from 'wagmi'
+import { useState, useEffect, Fragment, SetStateAction } from 'react'
+import { formatUnits, zeroAddress } from 'viem'
+import { useAccount } from 'wagmi'
 
-import { ERC_20_CONTRACT_ABI } from '../../../constants/abi'
 import { SelectPaymentSettings } from '../../../contexts'
-import { useClearCachedBalances, useSelectPaymentModal } from '../../../hooks'
+import { useClearCachedBalances } from '../../../hooks'
 
 import { CryptoOption } from './CryptoOption'
 
 interface PayWithCryptoProps {
   settings: SelectPaymentSettings
   disableButtons: boolean
-  setDisableButtons: React.Dispatch<React.SetStateAction<boolean>>
+  selectedCurrency: string | undefined
+  setSelectedCurrency: React.Dispatch<SetStateAction<string | undefined>>
+  isLoading: boolean
 }
 
-export const PayWithCrypto = ({ settings, disableButtons, setDisableButtons }: PayWithCryptoProps) => {
+export const PayWithCrypto = ({
+  settings,
+  disableButtons,
+  selectedCurrency,
+  setSelectedCurrency,
+  isLoading
+}: PayWithCryptoProps) => {
   const { enableSwapPayments = true, enableMainCurrencyPayment = true } = settings
-  const [selectedCurrency, setSelectedCurrency] = useState<string>()
 
-  const {
-    chain,
-    currencyAddress,
-    targetContractAddress,
-    price,
-    txData,
-    transactionConfirmations = TRANSACTION_CONFIRMATIONS_DEFAULT,
-    onSuccess = () => {},
-    onError = () => {}
-  } = settings
-  const { address: userAddress, connector } = useAccount()
-  const { data: walletClient } = useWalletClient()
-  const publicClient = usePublicClient()
+  const { chain, currencyAddress, price } = settings
+  const { address: userAddress } = useAccount()
   const { clearCachedBalances } = useClearCachedBalances()
-  const { closeSelectPaymentModal } = useSelectPaymentModal()
   const network = findSupportedNetwork(chain)
   const chainId = network?.chainId || 137
-
-  const {
-    data: allowanceData,
-    isLoading: allowanceIsLoading,
-    refetch: refechAllowance
-  } = useReadContract({
-    abi: ERC_20_CONTRACT_ABI,
-    functionName: 'allowance',
-    chainId: chainId,
-    address: currencyAddress as Hex,
-    args: [userAddress, targetContractAddress],
-    query: {
-      enabled: !!userAddress
-    }
-  })
 
   const { data: currencyBalanceData, isLoading: currencyBalanceIsLoading } = useBalances({
     chainIds: [chainId],
@@ -89,30 +58,7 @@ export const PayWithCrypto = ({ settings, disableButtons, setDisableButtons }: P
     { disabled: !enableSwapPayments }
   )
 
-  const disableSwapQuote = !selectedCurrency || compareAddress(settings.currencyAddress, selectedCurrency || '')
-
-  const { data: swapQuote, isLoading: isLoadingSwapQuote } = useSwapQuote(
-    {
-      userAddress: userAddress ?? '',
-      buyCurrencyAddress: settings?.currencyAddress,
-      buyAmount: price,
-      chainId: chainId,
-      sellCurrencyAddress: selectedCurrency || '',
-      includeApprove: true
-    },
-    {
-      disabled: !selectedCurrency
-    }
-  )
-
-  const mainToken = [
-    {
-      chainId,
-      contractAddress: currencyAddress
-    }
-  ]
-
-  const isLoading = allowanceIsLoading || currencyBalanceIsLoading || isLoadingCurrencyInfo
+  const isLoadingOptions = currencyBalanceIsLoading || isLoadingCurrencyInfo || isLoading
 
   const swapsIsLoading = swapPricesIsLoading
 
@@ -141,8 +87,6 @@ export const PayWithCrypto = ({ settings, disableButtons, setDisableButtons }: P
   ]
 
   const priceFormatted = formatUnits(BigInt(price), currencyInfoData?.decimals || 0)
-  const isNativeToken = compareAddress(currencyAddress, zeroAddress)
-  const isApproved: boolean = (allowanceData as bigint) >= BigInt(price) || isNativeToken
 
   const balanceInfo = currencyBalanceData?.find(balanceData => compareAddress(currencyAddress, balanceData.contractAddress))
 
@@ -155,157 +99,6 @@ export const PayWithCrypto = ({ settings, disableButtons, setDisableButtons }: P
   useEffect(() => {
     clearCachedBalances()
   }, [])
-
-  const onPurchaseMainCurrency = async () => {
-    if (!walletClient || !userAddress || !publicClient || !userAddress || !connector) {
-      return
-    }
-
-    setDisableButtons(true)
-
-    try {
-      const walletClientChainId = await walletClient.getChainId()
-      if (walletClientChainId !== chainId) {
-        await walletClient.switchChain({ id: chainId })
-      }
-
-      const approveTxData = encodeFunctionData({
-        abi: ERC_20_CONTRACT_ABI,
-        functionName: 'approve',
-        args: [targetContractAddress, price]
-      })
-
-      const transactions = [
-        ...(isApproved
-          ? []
-          : [
-              {
-                to: currencyAddress as Hex,
-                data: approveTxData,
-                chainId
-              }
-            ]),
-        {
-          to: targetContractAddress as Hex,
-          data: txData,
-          chainId,
-          ...(isNativeToken
-            ? {
-                value: BigInt(settings.price)
-              }
-            : {})
-        }
-      ]
-
-      const txHash = await sendTransactions({
-        chainId,
-        senderAddress: userAddress,
-        publicClient,
-        walletClient,
-        connector,
-        transactions,
-        transactionConfirmations
-      })
-
-      closeSelectPaymentModal()
-      refechAllowance()
-      clearCachedBalances()
-      onSuccess(txHash)
-    } catch (e) {
-      console.error('Failed to purchase...', e)
-      onError(e as Error)
-    }
-
-    setDisableButtons(false)
-  }
-
-  const onClickPurchaseSwap = async (swapPrice: SwapPricesWithCurrencyInfo) => {
-    if (!walletClient || !userAddress || !publicClient || !userAddress || !connector || !swapQuote) {
-      return
-    }
-
-    setDisableButtons(true)
-
-    try {
-      const walletClientChainId = await walletClient.getChainId()
-      if (walletClientChainId !== chainId) {
-        await walletClient.switchChain({ id: chainId })
-      }
-
-      const approveTxData = encodeFunctionData({
-        abi: ERC_20_CONTRACT_ABI,
-        functionName: 'approve',
-        args: [targetContractAddress, price]
-      })
-
-      const isSwapNativeToken = compareAddress(NATIVE_TOKEN_ADDRESS_0X, swapPrice.price.currencyAddress)
-
-      const transactions = [
-        // Swap quote optional approve step
-        ...(swapQuote?.approveData && !isSwapNativeToken
-          ? [
-              {
-                to: swapPrice.price.currencyAddress as Hex,
-                data: swapQuote.approveData as Hex,
-                chain: chainId
-              }
-            ]
-          : []),
-        // Swap quote tx
-        {
-          to: swapQuote.to as Hex,
-          data: swapQuote.transactionData as Hex,
-          chain: chainId,
-          ...(isSwapNativeToken
-            ? {
-                value: BigInt(swapQuote.transactionValue)
-              }
-            : {})
-        },
-        // Actual transaction optional approve step
-        ...(isApproved || isNativeToken
-          ? []
-          : [
-              {
-                to: currencyAddress as Hex,
-                data: approveTxData as Hex,
-                chainId: chainId
-              }
-            ]),
-        // transaction on the contract
-        {
-          to: targetContractAddress as Hex,
-          data: txData as Hex,
-          chainId,
-          ...(isNativeToken
-            ? {
-                value: BigInt(settings.price)
-              }
-            : {})
-        }
-      ]
-
-      const txHash = await sendTransactions({
-        chainId,
-        senderAddress: userAddress,
-        publicClient,
-        walletClient,
-        connector,
-        transactions,
-        transactionConfirmations
-      })
-
-      closeSelectPaymentModal()
-      refechAllowance()
-      clearCachedBalances()
-      onSuccess(txHash)
-    } catch (e) {
-      console.error('Failed to purchase...', e)
-      onError(e as Error)
-    }
-
-    setDisableButtons(false)
-  }
 
   const Options = () => {
     return (
@@ -367,17 +160,6 @@ export const PayWithCrypto = ({ settings, disableButtons, setDisableButtons }: P
     )
   }
 
-  const onClickPurchase = () => {
-    if (compareAddress(selectedCurrency || '', currencyAddress)) {
-      onPurchaseMainCurrency()
-    } else {
-      const foundSwap = swapPrices?.find(price => price.info?.address === selectedCurrency)
-      if (foundSwap) {
-        onClickPurchaseSwap(foundSwap)
-      }
-    }
-  }
-
   const gutterHeight = 8
   const optionHeight = 72
   const displayedOptionsAmount = Math.min(coins.length, 3)
@@ -388,13 +170,13 @@ export const PayWithCrypto = ({ settings, disableButtons, setDisableButtons }: P
 
   return (
     <Box width="full">
-      <Box marginTop="3">
+      <Box>
         <Text variant="small" fontWeight="medium" color="white">
           Pay with crypto
         </Text>
       </Box>
-      <Scroll paddingY="3" style={{ height: viewheight }}>
-        {isLoading ? (
+      <Scroll paddingY="3" style={{ height: viewheight, marginBottom: '-12px' }}>
+        {isLoadingOptions ? (
           <Box width="full" paddingTop="5" justifyContent="center" alignItems="center">
             <Spinner />
           </Box>
@@ -402,15 +184,6 @@ export const PayWithCrypto = ({ settings, disableButtons, setDisableButtons }: P
           <Options />
         )}
       </Scroll>
-      <Button
-        onClick={onClickPurchase}
-        disabled={isLoading || disableButtons || !selectedCurrency || (!disableSwapQuote && isLoadingSwapQuote)}
-        marginTop="2"
-        shape="square"
-        variant="primary"
-        width="full"
-        label="Complete Purchase"
-      />
     </Box>
   )
 }
